fix test

+++ server/src/routes/todos.controller.js
@@ -1,4 +1,4 @@
-const { getAllTodos, saveTodo, changeTodo, deleteTodo } = require('../models/todos.model');
+const { getAllTodos, saveTodo, changeTodo, deleteTodo, findTodo } = require('../models/todos.model');
 
 const getTodos = async (req, res) => {
   try {
@@ -16,9 +16,9 @@ const getTodos = async (req, res) => {
 const addTodo = async (req, res) => {
   const newTodo = req.body;
 
-  if (!newTodo.name) {
+  if (!newTodo.name || !newTodo.id) {
     return res.status(400).json({
-      error: "Missing required name property"
+      error: "missing required property"
     })
   }
   try {
@@ -36,6 +36,17 @@ const addTodo = async (req, res) => {
 
 const updateTodo = async (req, res) => {
   const newTodo = req.body;
+  if (!newTodo.name || !newTodo.id) {
+    return res.status(400).json({
+      error: "missing required property"
+    })
+  }
+  const isTodoExist = await findTodo(newTodo.id);
+  if (!isTodoExist) {
+    return res.status(404).json({
+      error: "todo not found"
+    })
+  }
   try {
     await changeTodo(newTodo);
     return res.status(200).json({
@@ -43,14 +54,17 @@ const updateTodo = async (req, res) => {
     });
   } catch (error) {
     console.error(error);
-    return res.status(404).json({
-      error: "todo not found"
-    })
   }
 };
 
 const removeTodo = async (req, res) => {
   const { id } = req.params;
+  const isTodoExist = await findTodo(id);
+  if (!isTodoExist) {
+    return res.status(404).json({
+      error: "todo not found"
+    })
+  }
   try {
     await deleteTodo(id);
     return res.status(200).json({
@@ -58,8 +72,8 @@ const removeTodo = async (req, res) => {
     });
   } catch (error) {
     console.error(error);
-    return res.status(404).json({
-      error: "todo not found"
+    return res.status(400).json({
+      error: "error with delete todo"
     })
   }
 };
